+++ src/initial_condition/initial_condition.h
@@ -6,10 +6,10 @@
 class InitialCondition
 {
     private:
-        std::function<std::complex<double>(double, double)> initial_condition_function;
+        std::function<std::complex<float>(float, float)> initial_condition_function;
     public:
         InitialCondition() = default;
-        InitialCondition(std::function<std::complex<double>(double, double)> initial_condition_function);
+        InitialCondition(std::function<std::complex<float>(float, float)> initial_condition_function);
         void assign_to_domain(BaseDomain *domain);
         
 };
